backend/requirements.txt
flask
flask-cors
torch
torchvision
Pillow
opencv-python
ultralytics